+++ generate/src/bin/generate_from_types/docs_config.rs
