+++ generate/src/bin/generate_from_types/document.rs
