+++ generate/src/bin/generate_from_types/markdown/properties.rs
