+++ generate/src/bin/generate_from_types/schema_parser/utils.rs
